moved WeatherStation Observer pattern exercise
