--- Design-Patterns/Observer-Pattern/Display.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Observer_Pattern
-{
-    interface Display
-    {
-        void Display();
-    }
-}
