--- Design-Patterns/Observer-Pattern/Subject.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Observer_Pattern
-{
-    interface Subject
-    {
-        void RegisterObserver(Observer observer);
-
-        void RemoveObserver(Observer observer);
-
-        void NotifyObservers();
-    }
-}
