+++ Observer/Weather-Station/Program.cs
@@ -4,7 +4,17 @@
     {
         static void Main(string[] args)
         {
-            Console.WriteLine("Hello, World!");
+            List<Display> displays = new List<Display>();
+
+            WeatherData station = new WeatherData();
+
+            displays.Add(new CurrentConditionsDisplay(station));
+            displays.Add(new ForecastDisplay(station));
+            displays.Add(new StatisticsDisplay(station));
+
+            station.SetMeasurements(10, 10, 10);
+            station.SetMeasurements(20, 10, 29);
+            station.SetMeasurements(10, 30, 10);
         }
     }
 }
